Create Core.php, put a lot of helper methods here instead.
